place_water speedup
using torch for the function + using single argsort instead of repeated useage of argmax (number of argmax call = number of placed water)

+++ GWCNN/gpu/btorch_gpu.py
@@ -217,7 +217,7 @@ def build_cuda(vecs,fgs,grid=0.5,n_grid=48,weight=25.0,multi=1.5):
 def build_cuda_rect(vecs,fgs,grid=0.5,n_grid_arr=[48,48,48],weight=25.0,multi=1.5):
     #prepare model
     device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-    ng = torch.tensor(torch.IntTensor(n_grid_arr), device=device ,requires_grad=False, dtype=torch.int32)
+    ng = torch.tensor(n_grid_arr, device=device, requires_grad=False, dtype=torch.int32)
     ng = ng.to(device=device) #gpu
     model_pro = Model_pro(grid,ng)
     model_pro.to(device)
